refactor: remake pagination-nav-item, shortened pagination component
