+++ src/shared/ui/pagination/pagination.tsx
@@ -2,15 +2,13 @@ import { type ComponentProps, useMemo } from 'react'
 
 import { clsx } from 'clsx'
 
-import { Button } from '@/shared/ui/button'
-import { ChevronRightIcon } from '@/shared/ui/icons'
 import { PaginationNavItem } from '@/shared/ui/pagination/pagination-nav-item'
 import { Select } from '@/shared/ui/select'
 import { Typography } from '@/shared/ui/typography'
 
 import styles from './pagination.module.scss'
 
-import { DOTS, usePagination } from './use-pagination'
+import { usePagination } from './use-pagination'
 
 type PaginationProps = {
    onPageChange: (page: number) => void
@@ -42,73 +40,18 @@ export const Pagination = ({
       [pageSizeOptions]
    )
 
-   const handleNextPage = () => {
-      if (currentPage < lastPage) {
-         onPageChange(currentPage + 1)
-      }
-   }
-
-   const handlePrevPage = () => {
-      if (currentPage > 1) {
-         onPageChange(currentPage - 1)
-      }
-   }
-
    const handlePageSizeChange = (value: string) => {
-      onPageSizeChange?.(Number(value))
+      onPageSizeChange(Number(value))
    }
-
-   const lastPage = paginationRange[paginationRange.length - 1] as number
-
-   const isPrevDisabled = currentPage === 1
-   const isNextDisabled = currentPage === lastPage
 
    return (
       <div className={clsx(styles.paginationContainer, className)}>
          <div className={styles.paginationWrapper}>
-            <ul className={styles.pagination}>
-               <li className={clsx(styles.arrow, styles.leftArrow)}>
-                  <Button
-                     disabled={isPrevDisabled}
-                     variant={'ghost'}
-                     size={'icon'}
-                     onClick={handlePrevPage}
-                  >
-                     <ChevronRightIcon width={16} height={16} />
-                  </Button>
-               </li>
-               {paginationRange.map((pageNum, index) => {
-                  const isCurrent = pageNum === currentPage
-
-                  if (pageNum === DOTS) {
-                     return (
-                        <li key={`dots-${index}`} className={styles.paginationDots}>
-                           &#8230;
-                        </li>
-                     )
-                  }
-
-                  return (
-                     <PaginationNavItem
-                        key={`page-${pageNum}`}
-                        className={clsx(styles.paginationItem, isCurrent && styles.selected)}
-                        onClick={() => onPageChange(pageNum)}
-                     >
-                        {pageNum}
-                     </PaginationNavItem>
-                  )
-               })}
-               <li className={clsx(styles.arrow, styles.rightArrow)}>
-                  <Button
-                     disabled={isNextDisabled}
-                     variant={'ghost'}
-                     size={'icon'}
-                     onClick={handleNextPage}
-                  >
-                     <ChevronRightIcon width={16} height={16} />
-                  </Button>
-               </li>
-            </ul>
+            <PaginationNavItem
+               paginationRange={paginationRange}
+               currentPage={currentPage}
+               onPageChange={onPageChange}
+            />
          </div>
          <div className={styles.itemsCountWrapper}>
             <Typography variant={'body2'}>Показать</Typography>
